refactor: fixes error hierarchy
