+++ src/core/errors/url-error.ts
@@ -17,21 +17,3 @@ export class UrlNotFoundError extends DomainError {
     super(`URL with short code "${shortCode}" not found`);
   }
 }
-
-export class UrlCreationFailedError extends Error {
-  constructor(
-    originalUrl: string,
-    public readonly cause?: Error,
-  ) {
-    super(`Failed to create short URL for "${originalUrl}"`);
-    this.name = 'UrlCreationFailedError';
-
-    // Garantir que o protótipo seja mantido
-    Object.setPrototypeOf(this, UrlCreationFailedError.prototype);
-
-    // Capturar stack trace
-    if (Error.captureStackTrace) {
-      Error.captureStackTrace(this, this.constructor);
-    }
-  }
-}
